cleanup: remove group resources view
