+++ src/components/charactersResourcesView/charactersResourcesView.schema.yaml
@@ -1,0 +1,25 @@
+componentName: rvn-characters-resources-view
+propsSchema:
+  type: object
+  properties:
+    navTitle:
+      type: string
+      description: Optional title to display in the header
+    groups:
+      type: array
+      description: Array of character groups to display
+    selectedItemId:
+      type: string
+      description: ID of the currently selected character
+    searchQuery:
+      type: string
+      description: Current search query
+    searchPlaceholder:
+      type: string
+      description: Placeholder text for search input
+    emptyMessage:
+      type: string
+      description: Message to show when no items match
+    addText:
+      type: string
+      description: Text to display in add affordances
