+++ src/components/charactersResourcesView/charactersResourcesView.store.js
@@ -1,0 +1,74 @@
+export const createInitialState = () => ({
+  collapsedIds: [],
+  dropdownMenu: {
+    isOpen: false,
+    x: 0,
+    y: 0,
+    targetItemId: undefined,
+    items: [],
+  },
+});
+
+export const toggleGroupCollapse = ({ state }, { groupId } = {}) => {
+  const index = state.collapsedIds.indexOf(groupId);
+  if (index > -1) {
+    state.collapsedIds.splice(index, 1);
+    return;
+  }
+
+  state.collapsedIds.push(groupId);
+};
+
+export const showContextMenu = ({ state }, { itemId, x, y } = {}) => {
+  state.dropdownMenu.isOpen = true;
+  state.dropdownMenu.x = x;
+  state.dropdownMenu.y = y;
+  state.dropdownMenu.targetItemId = itemId;
+  state.dropdownMenu.items = [
+    { label: "Delete", type: "item", value: "delete-item" },
+  ];
+};
+
+export const hideContextMenu = ({ state }, _payload = {}) => {
+  state.dropdownMenu.isOpen = false;
+  state.dropdownMenu.x = 0;
+  state.dropdownMenu.y = 0;
+  state.dropdownMenu.targetItemId = undefined;
+  state.dropdownMenu.items = [];
+};
+
+export const selectDropdownMenu = ({ state }) => state.dropdownMenu;
+
+export const selectViewData = ({ state, props }) => {
+  const groups = (props.groups ?? []).map((group) => {
+    const isCollapsed = state.collapsedIds.includes(group.id);
+    const children = isCollapsed ? [] : (group.children ?? []);
+
+    return {
+      ...group,
+      isCollapsed,
+      children: children.map((item) => {
+        const isSelected = item.id === props.selectedItemId;
+
+        return {
+          ...item,
+          itemBorderColor: isSelected ? "pr" : "bo",
+          itemHoverBorderColor: isSelected ? "pr" : "ac",
+        };
+      }),
+    };
+  });
+
+  return {
+    navTitle: props.navTitle,
+    groups,
+    selectedItemId: props.selectedItemId,
+    searchQuery: props.searchQuery ?? "",
+    searchPlaceholder: props.searchPlaceholder ?? "Search...",
+    emptyMessage:
+      props.emptyMessage ??
+      `No characters found matching "${props.searchQuery ?? ""}"`,
+    addText: props.addText ?? "Add Character",
+    dropdownMenu: state.dropdownMenu,
+  };
+};
