--- src/components/groupResourcesView/groupResourcesView.schema.yaml
@@ -1,58 +1,0 @@
-componentName: rvn-group-resources-view
-propsSchema:
-  type: object
-  properties:
-    resourceType:
-      type: string
-      description: Type of resource (fonts, images, audio, etc.)
-    flatGroups:
-      type: array
-      description: Array of resource groups to display (pre-filtered and processed by parent)
-    selectedItemId:
-      type: string
-      description: ID of currently selected item
-    searchQuery:
-      type: string
-      description: Current search query (controlled by parent)
-    collapsedIds:
-      type: array
-      description: Array of collapsed group IDs (controlled by parent)
-    navTitle:
-      type: string
-      description: Optional title to display in the header
-    uploadText:
-      type: string
-      description: Text to display in upload zone
-    acceptedFileTypes:
-      type: array
-      description: Array of accepted file extensions
-    searchPlaceholder:
-      type: string
-      description: Placeholder text for search input
-    emptyMessage:
-      type: string
-      description: Message to show when no items found
-    imageHeight:
-      type: number
-      description: Height for image items (for zoom control)
-    maxWidth:
-      type: number
-      description: Max width for image items (for zoom control)
-    zoomLevel:
-      type: number
-      description: Current zoom level for zoom controls
-    showZoomControls:
-      type: boolean
-      description: Whether to show zoom controls
-    fullWidthItem:
-      type: boolean
-      description: Whether each item should stretch to full available width
-    backUrl:
-      type: string
-      description: URL to navigate back to when back button is clicked
-    itemProperties:
-      type: object
-      description: Optional field/resource properties map for item rendering
-    items:
-      type: object
-      description: Optional repository items map keyed by resource id
