+++ src/components/typographyResourcesView/typographyResourcesView.view.yaml
@@ -1,0 +1,68 @@
+refs:
+  searchInput:
+    eventListeners:
+      value-input:
+        handler: handleSearchInput
+  contextMenu:
+    eventListeners:
+      close:
+        handler: handleCloseContextMenu
+      item-click:
+        handler: handleContextMenuClickItem
+  addBtn*:
+    eventListeners:
+      click:
+        handler: handleAddButtonClick
+  groupToggle*:
+    eventListeners:
+      click:
+        handler: handleGroupClick
+  item*:
+    eventListeners:
+      click:
+        handler: handleItemClick
+      dblclick:
+        handler: handleItemDoubleClick
+      contextmenu:
+        handler: handleItemContextMenu
+template:
+  - 'rtgl-view w=f h=100vh g=lg style="min-width: 0;"':
+      - 'rtgl-view h=48 w=f bgc=bg bwb=xs ph=md av=c style="position: sticky; top: 0px; z-index: 1000;"':
+          - rtgl-view w=f d=h av=c g=md wrap:
+              - rtgl-view av=c g=sm d=h w=1fg:
+                  - $if navTitle:
+                      - rtgl-text c=mu-fg: ${navTitle}
+              - rtgl-input#searchInput s=sm placeholder=${searchPlaceholder} value=${searchQuery} h=1fg: null
+      - 'rtgl-view h=1fg w=f sv style="min-width: 0;"':
+          - $if groups.length > 0:
+              - $for group, gIndex in groups:
+                  - rtgl-view w=f ph=md:
+                      - 'rtgl-view d=h w=f av=c style="position: sticky; top: 0px;"':
+                          - rtgl-view#groupToggleRef${gIndex} data-group-id=${group.id} d=h av=c g=md cur=pointer:
+                              - $if group.isCollapsed:
+                                  - rtgl-svg wh=16 svg=folderArrowRight: null
+                                $else:
+                                  - rtgl-svg wh=16 svg=folderArrowDown: null
+                              - rtgl-svg wh=16 svg=folder: null
+                              - rtgl-text: ${group.fullLabel}
+                          - rtgl-view w=1fg: null
+                          - rtgl-button#addBtnHeaderRef${gIndex} data-group-id=${group.id} s=sm w=1fg pre=plus: ${addText}
+                      - $if !group.isCollapsed:
+                          - 'rtgl-view w=f d=h ph=sm pt=md pb=sm mb=lg wrap g=md style="min-width: 0;"':
+                              - $if group.hasChildren:
+                                  - $for item, itemIndex in group.children:
+                                      - rtgl-view#itemRef${gIndex}x${itemIndex} data-item-id=${item.id} w=360 cur=pointer bw=xs bc=${item.itemBorderColor} h-bc=${item.itemHoverBorderColor} br=md:
+                                          - rtgl-view p=md g=md w=f br=md:
+                                              - rvn-font-preview fontFamily=${item.fontStyle} previewText="${item.previewText}" fileId=${item.fontFileId} fontSize=${item.fontSize} lineHeight=${item.lineHeight} color=${item.color} fontWeight=${item.fontWeight} width=${previewWidth} height=96 showBorder=false: null
+                                              - rtgl-text s=xs c=mu-fg ellipsis w=f: ${item.name}
+                                $else:
+                                  - rtgl-view#addBtnEmptyRef${gIndex} data-group-id=${group.id} w=f h=150 av=c ah=c g=md p=lg bw=xs bc=bo br=lg cur=pointer:
+                                      - rtgl-svg svg=plus wh=24 c=pr: null
+                                      - rtgl-text s=md c=pr fw=500: ${addText}
+            $elif searchQuery:
+              - rtgl-view w=f h=400 av=c ah=c:
+                  - rtgl-text s=lg c=mu-fg: ${emptyMessage}
+            $else:
+              - rtgl-view w=f h=f av=c ah=c:
+                  - rtgl-text s=lg c=mu-fg: No data
+  - rtgl-dropdown-menu#contextMenu :items=dropdownMenu.items ?open=${dropdownMenu.isOpen} x=${dropdownMenu.x} y=${dropdownMenu.y}: null
