+++ src/pages/characters/characters.store.js
@@ -34,7 +34,6 @@ export const createInitialState = () => ({
   charactersData: { tree: [], items: {} },
   selectedItemId: null,
   searchQuery: "",
-  collapsedIds: [],
   isDialogOpen: false,
   targetGroupId: null,
   avatarFileId: null,
@@ -98,15 +97,6 @@ export const setSelectedItemId = ({ state }, { itemId } = {}) => {
 
 export const setSearchQuery = ({ state }, { query } = {}) => {
   state.searchQuery = query;
-};
-
-export const toggleGroupCollapse = ({ state }, { groupId } = {}) => {
-  const index = state.collapsedIds.indexOf(groupId);
-  if (index > -1) {
-    state.collapsedIds.splice(index, 1);
-  } else {
-    state.collapsedIds.push(groupId);
-  }
 };
 
 export const setTargetGroupId = ({ state }, { groupId } = {}) => {
@@ -217,20 +207,7 @@ export const selectViewData = ({ state }) => {
       .filter(Boolean);
   }
 
-  // Apply collapsed state and selection styling
-  const flatGroups = filteredGroups.map((group) => ({
-    ...group,
-    isCollapsed: state.collapsedIds.includes(group.id),
-    children: state.collapsedIds.includes(group.id)
-      ? []
-      : (group.children || []).map((item) => ({
-          ...item,
-          selectedStyle:
-            item.id === state.selectedItemId
-              ? "outline: 2px solid var(--color-pr); outline-offset: 2px;"
-              : "",
-        })),
-  }));
+  const flatGroups = filteredGroups;
 
   // Get edit item details
   const editItem = state.editItemId
