+++ src/pages/characters/characters.view.yaml
@@ -9,7 +9,7 @@ refs:
         handler: handleFileExplorerSelectionChanged
       dblclick-item:
         handler: handleSpritesButtonClick
-  groupview:
+  charactersView:
     eventListeners:
       item-click:
         handler: handleCharacterItemClick
@@ -21,8 +21,6 @@ refs:
         handler: handleAddCharacterClick
       search-input:
         handler: handleSearchInput
-      group-toggle:
-        handler: handleGroupToggle
       item-delete:
         handler: handleItemDelete
   addCharacterDialog:
@@ -69,7 +67,7 @@ template:
                   - rvn-resource-types :resourceCategory=resourceCategory :selectedResourceId=selectedResourceId: null
                   - rvn-base-file-explorer#fileExplorer :items=flatItems shrinkable dropdown-menu allowDrag :folderContextMenuItems=folderContextMenuItems :itemContextMenuItems=itemContextMenuItems :emptyContextMenuItems=emptyContextMenuItems: null
           - rtgl-view w=1fg h=f:
-              - rvn-group-resources-view#groupview :navTitle=title full-width-item :resourceType=resourceType :flatGroups=flatGroups :selectedItemId=selectedItemId :searchQuery=searchQuery show-zoom-controls=false: null
+              - rvn-characters-resources-view#charactersView :navTitle=title :groups=flatGroups :selectedItemId=selectedItemId :searchQuery=searchQuery: null
           - rvn-resizable-panel panel-type=detail-panel w=270 min-w=200 max-w=500 resize-side="left":
               - rtgl-view wh=f slot="content":
                   - $if selectedItemId:
